+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/session/DefaultMeetingSession.kt
@@ -39,7 +39,8 @@ class DefaultMeetingSession(
         val audioClientObserver =
             DefaultAudioClientObserver(
                 logger,
-                metricsCollector
+                metricsCollector,
+                configuration
             )
 
         val audioClient =